requirements.txt
# AWJ Calculator Pro - Python Dependencies
# For PythonAnywhere and local development

# Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Scientific Computing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3

# Machine Learning & AI
scikit-learn==1.3.2
tensorflow==2.15.0  # nebo tensorflow-cpu pro PythonAnywhere
torch==2.1.1  # PyTorch pro advanced AI

# Data Visualization
matplotlib==3.8.2
plotly==5.18.0

# API & Web
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0

# Celery for async tasks (optional)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
